+++ src/config.lua
@@ -2,4 +2,5 @@ return {
   version = 0;
   enabled = true;
   AlwaysUnlockAfterPurge = false,
+  AutoUnlockAndUpgradeCards = false,
 }